refactor: change component names
