+++ client/src/components/VideoCall.jsx
@@ -15,7 +15,6 @@ export default function VideoCall() {
     name,
     video,
     setReceivingCall,
-    // messagingComponent,
   } = useVideoCall();
 
   return (
@@ -86,11 +85,6 @@ export default function VideoCall() {
             </div>
           </div>
         ) : null}
-
-        {/* {messagingComponent ? (
-          // Render the messaging component here
-          <div>Messaging Component</div>
-        ) : null} */}
       </div>
     </>
   );
